[feature]: optimize code
